1.自定义stater初版完成 2.swagger3.0-keycloak-starter:user服务验证通过;knife4j接口文档,用户名密码验证模式待实现

+++ service-message-parent/service-message-client/src/main/java/com/gsh/springcloud/message/request/MessageReq.java
@@ -16,11 +16,11 @@ public class MessageReq {
   /**
    * 用户id
    */
-  private Integer userId;
+  private Long userId;
   /**
    * 积分变化值
    */
-  private Integer integral;
+  private Double integral;
   /**
    * 订单id
    */
